django-backend/requirements.txt
asgiref==3.5.2
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
Pillow==9.2.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
mysqlclient==2.1.1
gunicorn>=21,<22
whitenoise>=6.6.0,<7
django-filter>=23.5,<24
coloredlogs>=15.0.1,<16
flake8>=7.0.0,<8.0.0